+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## 0.2.2 (2019-10-17)
+* Added: Parameter for onlinePath prefix to the files.
+## 0.2.0 (2019-05-21)
+* Updated: rollup is updated to v1.1.0.
+* updated: onwrite hook is replaced by writeBundle.
+## 0.1.5 (2019-03-21)
+* added: support for HTML code string as template.
 ## 0.1.3 (2018-10-10)
 * fixed: sourcemap is not hashed.